Virt module refactoring
This commit refactors the virt module. This is a long overdue refactor,
as our virt module ended-up as a single file holding all of the code
required for privileged instructions emulation.

With the refactor the file is now split in sub-modules with more limited
responsibilities:
- CSR emulation
- Privileged instruction emulation
- And world switches
